+++ spec/classes/install/configuration/mvn_url_spec.rb
@@ -9,7 +9,7 @@ describe 'karaf::install::configuration::mvn_url' do
       let(:pre_condition) { 'service { "karaf": }' }
       let(:params) do
         {
-          etc_dir: '/opt/karaf/work/etc/',
+          etc_dir: '/opt/karaf/apache-karaf-4.4.11/etc/',
           service_name: 'karaf',
           service_user_name: 'karaf',
           service_group_name: 'karaf',
@@ -18,7 +18,7 @@ describe 'karaf::install::configuration::mvn_url' do
       end
 
       it { is_expected.to compile.with_all_deps }
-      it { is_expected.to contain_file('/opt/karaf/work/etc/org.ops4j.pax.url.mvn.cfg').with(owner: 'karaf', group: 'karaf') }
+      it { is_expected.to contain_file('/opt/karaf/apache-karaf-4.4.11/etc/org.ops4j.pax.url.mvn.cfg').with(owner: 'karaf', group: 'karaf') }
     end
   end
 end