+++ spec/classes/install/configuration/setenv_spec.rb
@@ -9,7 +9,7 @@ describe 'karaf::install::configuration::setenv' do
       let(:pre_condition) { 'service { "karaf": }' }
       let(:params) do
         {
-          bin_dir: '/opt/karaf/apache-karaf-4.4.3/bin/',
+          bin_dir: '/opt/karaf/apache-karaf-4.4.11/bin/',
           service_name: 'karaf',
           java_home: '/usr/lib/jvm/java',
           default_env_vars: { 'JAVA_OPTS' => '-Xmx512m' },
@@ -17,8 +17,8 @@ describe 'karaf::install::configuration::setenv' do
       end
 
       it { is_expected.to compile.with_all_deps }
-      it { is_expected.to contain_file_line('/opt/karaf/apache-karaf-4.4.3/bin/setenv-JAVA_HOME').with(path: '/opt/karaf/apache-karaf-4.4.3/bin/setenv', line: 'export JAVA_HOME="/usr/lib/jvm/java"', match: '^export JAVA_HOME=') }
-      it { is_expected.to contain_file_line('/opt/karaf/apache-karaf-4.4.3/bin/setenv-JAVA_OPTS').with(path: '/opt/karaf/apache-karaf-4.4.3/bin/setenv', line: 'export JAVA_OPTS=-Xmx512m', match: '^export JAVA_OPTS=') }
+      it { is_expected.to contain_file_line('/opt/karaf/apache-karaf-4.4.11/bin/setenv-JAVA_HOME').with(path: '/opt/karaf/apache-karaf-4.4.11/bin/setenv', line: 'export JAVA_HOME="/usr/lib/jvm/java"', match: '^export JAVA_HOME=') }
+      it { is_expected.to contain_file_line('/opt/karaf/apache-karaf-4.4.11/bin/setenv-JAVA_OPTS').with(path: '/opt/karaf/apache-karaf-4.4.11/bin/setenv', line: 'export JAVA_OPTS=-Xmx512m', match: '^export JAVA_OPTS=') }
     end
   end
 end