+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## Release 1.1.0
+
+**Fixed**
+- no shared etc & instances dir over different Karaf versions, after upgrade instances need to be build from scratch
+
 ## Release 1.0.1
 
 **Features**